+++ ethergo/submitter/submitter.go
@@ -4,7 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/ethereum/go-ethereum/params"
 	"github.com/puzpuzpuz/xsync/v2"
+	"github.com/synapsecns/sanguine/ethergo/chain/gas/backend"
+	"github.com/synapsecns/sanguine/ethergo/chain/gas/london"
 	"math"
 	"math/big"
 	"reflect"
@@ -213,7 +216,7 @@ func (t *txSubmitterImpl) getNonce(parentCtx context.Context, chainID *big.Int, 
 
 func (t *txSubmitterImpl) storeTX(ctx context.Context, tx *types.Transaction, status db.Status) (err error) {
 	ctx, span := t.metrics.Tracer().Start(ctx, "submitter.StoreTX", trace.WithAttributes(
-		append(txToAttributes(tx), attribute.String("status", status.String()))...))
+		append(util.TxToAttributes(tx), attribute.String("status", status.String()))...))
 
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
@@ -271,7 +274,7 @@ func (t *txSubmitterImpl) SubmitTransaction(parentCtx context.Context, chainID *
 	}
 
 	// then we copy the transactor, this is the one we'll modify w/ no send.
-	transactor := copyTransactOpts(parentTransactor)
+	transactor := util.CopyTransactOpts(parentTransactor)
 
 	var locker mapmutex.Unlocker
 
@@ -362,7 +365,8 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 	if t.config.SupportsEIP1559(int(bigChainID.Uint64())) {
 		transactor.GasFeeCap = t.config.GetMaxGasPrice(chainID)
 
-		transactor.GasTipCap, err = client.SuggestGasTipCap(ctx)
+		// only use fallback if no previous tx
+		transactor.GasTipCap, err = t.SuggestGasTipCap(ctx, client, chainID, prevTx != nil)
 		if err != nil {
 			return fmt.Errorf("could not get gas tip cap: %w", err)
 		}
@@ -419,7 +423,7 @@ func (t *txSubmitterImpl) getGasBlock(ctx context.Context, chainClient client.EV
 		if ok {
 			span.AddEvent("could not get gas block; using cached value", trace.WithAttributes(
 				attribute.String("error", err.Error()),
-				attribute.String("blockNumber", bigPtrToString(gasBlock.Number)),
+				attribute.String("blockNumber", util.BigPtrToString(gasBlock.Number)),
 			))
 		} else {
 			return nil, fmt.Errorf("could not get gas block: %w", err)
@@ -466,5 +470,90 @@ func (t *txSubmitterImpl) getGasEstimate(ctx context.Context, chainClient client
 
 	return gasEstimate, nil
 }
+
+// forceNoFallbackIfZero is used to force no fallback if the tip cap is 0.
+var forceNoFallbackIfZero = false
+
+// TODO: test oracle fallback.
+func (t *txSubmitterImpl) SuggestGasTipCap(ctx context.Context, client client.EVM, chainID int, fallbackIfZero bool) (tipCap *big.Int, err error) {
+	tipCap, err = client.SuggestGasTipCap(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("could not get gas tip cap: %w", err)
+	}
+
+	// note: these ifs can be combined, but I've found the largic hard to parse if they are.
+	// if tip cap is not zero  return
+	if big.NewInt(0).Cmp(tipCap) != 0 {
+		return tipCap, nil
+	}
+
+	// past this point tip cap is 0.
+
+	// if we should force no fallback, return
+	// this is set for testing.
+	if forceNoFallbackIfZero {
+		return tipCap, nil
+	}
+
+	// if we shouldn't fallback, return
+	if !fallbackIfZero {
+		return tipCap, nil
+	}
+
+	blockHeight, err := client.BlockNumber(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("could not get block number: %w", err)
+	}
+
+	// if on the other hand, fee is 0 we should fallback to our pricer
+	oracleConfig := gas.GetConfig()
+	oracleConfig.MaxPrice = core.CopyBigInt(t.config.GetMaxGasPrice(chainID))
+
+	feeOracle := london.NewFeeOracle(wrappedLondonClient{client, chainID}, blockHeight, oracleConfig)
+	tipCap, err = feeOracle.SuggestTipCap(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("could not get tip cap: %w", err)
+	}
+	return tipCap, nil
+}
+
+type wrappedLondonClient struct {
+	client.EVM
+	chainID int
+}
+
+// ChainConfig is a fake chain config.
+// since wrapped client is only used for makeSigner
+// and on london, we just enable everything.
+func (w wrappedLondonClient) ChainConfig() *params.ChainConfig {
+	return &params.ChainConfig{
+		ChainID:                       big.NewInt(int64(w.chainID)),
+		HomesteadBlock:                big.NewInt(0),
+		DAOForkBlock:                  big.NewInt(0),
+		DAOForkSupport:                false,
+		EIP150Block:                   big.NewInt(0),
+		EIP150Hash:                    common.Hash{},
+		EIP155Block:                   big.NewInt(0),
+		EIP158Block:                   big.NewInt(0),
+		ByzantiumBlock:                big.NewInt(0),
+		ConstantinopleBlock:           big.NewInt(0),
+		PetersburgBlock:               big.NewInt(0),
+		IstanbulBlock:                 big.NewInt(0),
+		MuirGlacierBlock:              big.NewInt(0),
+		BerlinBlock:                   big.NewInt(0),
+		LondonBlock:                   big.NewInt(0),
+		ArrowGlacierBlock:             big.NewInt(0),
+		GrayGlacierBlock:              big.NewInt(0),
+		MergeNetsplitBlock:            big.NewInt(0),
+		ShanghaiBlock:                 big.NewInt(0),
+		CancunBlock:                   big.NewInt(0),
+		TerminalTotalDifficulty:       big.NewInt(0),
+		TerminalTotalDifficultyPassed: false,
+		Ethash:                        nil,
+		Clique:                        nil,
+	}
+}
+
+var _ backend.OracleBackendChain = &wrappedLondonClient{}
 
 var _ TransactionSubmitter = &txSubmitterImpl{}